+++ src/pages/OrderCheckoutPage.tsx
@@ -13,6 +13,7 @@ import loggedInUserState from "@/recoil/atoms/loggedInUserState";
 import Modal from "@/components/common/Modal";
 import { RequestCreateOrder, ShippingInfoType } from "@/types/orders";
 import ordersApi from "@/apis/services/orders";
+import LoadingSpinner from "@/components/common/LoadingSpinner";
 
 const OrderCheckoutPage = () => {
   const user = useRecoilValue(loggedInUserState);
@@ -23,7 +24,6 @@ const OrderCheckoutPage = () => {
   const [isStockModalOpen, setIsStockModalOpen] = useState(false);
   const [isShippingModalOpen, setIsShippingModalOpen] = useState(false);
   const [isOrderErrorModalOpen, setIsOrderErrorModalOpen] = useState(false);
-  const [isErrorModalOpen, setIsErrorModalOpen] = useState(false);
   const navigate = useNavigate();
   const location = useLocation();
 
@@ -73,8 +73,8 @@ const OrderCheckoutPage = () => {
   };
 
   const handleStockModalClose = () => {
-    navigate(-1);
     setIsStockModalOpen(false);
+    navigate(-1);
   };
 
   const handleShippingModalClose = () => {
@@ -82,8 +82,8 @@ const OrderCheckoutPage = () => {
   };
 
   const handleOrderErrorModalClose = () => {
-    navigate(-1);
     setIsOrderErrorModalOpen(false);
+    navigate(-1);
   };
 
   const handlePostOrder = () => {
@@ -101,19 +101,13 @@ const OrderCheckoutPage = () => {
     createOrder(data);
   };
 
-  const handleErrorModalClose = () => {
-    navigate(-1);
-    setIsErrorModalOpen(false);
-  };
-
   useEffect(() => {
-    if (!cartData || cartData.length <= 0) {
-      setIsErrorModalOpen(true);
-    }
     if (user) fetchAllCartItems();
   }, []);
 
-  if (user && cartData && cartData.length > 0) {
+  if (!user) return null;
+
+  if (cartData) {
     return (
       <OrderCheckoutPageLayer>
         <Modal isOpen={isOrderErrorModalOpen} message="주문에 실패했습니다.\n장바구니페이지로 돌아갑니다.">
@@ -143,13 +137,7 @@ const OrderCheckoutPage = () => {
       </OrderCheckoutPageLayer>
     );
   }
-  return (
-    <Modal isOpen={isErrorModalOpen} message="잘못된 접근입니다.">
-      <ButtonWrapper onClick={handleErrorModalClose}>
-        <Button value="확인" size="sm" variant="point"></Button>
-      </ButtonWrapper>
-    </Modal>
-  );
+  return <LoadingSpinner />;
 };
 
 export default OrderCheckoutPage;
